+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.5.0 - 2017-03-06
+------------------
+
+Cycle-Analyzer (infinite loop detection) added.
+Preliminary tests added
+
+
 0.4.5 - 2017-03-04
 ------------------